🔨 Change type of maxComments to int

+++ portfolio/src/main/java/com/google/sps/servlets/DataServlet.java
@@ -36,10 +36,10 @@ public class DataServlet extends HttpServlet {
 
   @Override
   public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-    long maxComments = Long.parseLong(request.getParameter("maxComments"));
+    int maxComments = Integer.parseInt(request.getParameter("maxComments"));
 
     // a value of -1 indicates that ALL comments are wanted
-    if (maxComments == -1) maxComments = Long.MAX_VALUE;
+    if (maxComments == -1) maxComments = Integer.MAX_VALUE;
 
     Query query = new Query("Comment").addSort("timestamp", SortDirection.DESCENDING);
